+++ landing/src/pages/Home.tsx
@@ -1,6 +1,7 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
 import Hero from "../components/Hero";
+import DownloadSection from "../components/DownloadSection";
 import Features from "../components/Features";
 import FAQ from "../components/FAQ";
 import About from "../components/About";
@@ -31,6 +32,7 @@ const Home = () => {
   return (
     <>
       <Hero />
+      <DownloadSection />
       <Features />
       <FAQ />
       <About />